sprite_animation: address codex P2s on #718
1. event_frames is now JSONC-authorable: []const u8 -> []const u16. The
   scene-bridge deserializer special-cases every []const u8 as a STRING
   (intern pool), so a []const u8 field could never be authored as a
   number array; []const u16 falls through to the generic slice->int
   branch. Verified by a direct []const u16 test and a full-component
   SpriteAnimation round-trip in deserializer_test.zig.

2. The tick now queues ONLY the event kinds the project declared, via a
   comptime SpriteAnimation.EventMask built from engineEventWanted. A game
   listening to only anim_frame no longer has its fixed PendingBuf crowded
   out / dropped by loop_end events from a multi-wrap tick. New
   advanceEventsMasked entry point; advanceEvents (all-kinds) kept for the
   #670 callers.

3. Loop-event traversal is bounded on dt spikes: repetition advances
   arithmetically (satAddU16Count, O(1)) and loop_end emission is capped at
   the buffer's remaining room instead of iterating once per wrap. A
   100k-wrap tick is O(bounded); repetition saturates at u16 max. ping_pong
   emitReversal bumps repetition regardless, emits only when loop_end wanted.

Tests + spec + fmt green on 0.16.0.

+++ test/sprite_animation_events_test.zig
@@ -94,7 +94,7 @@ test "tick: landing on an event frame emits engine__anim_frame with entity + ind
     var game = newGame(&recorder);
     defer game.deinit();
 
-    const marked = [_]u8{2};
+    const marked = [_]u16{2};
     const entity = game.createEntity();
     game.active_world.ecs_backend.addComponent(entity, Sprite{ .sprite_name = "a.png" });
     game.active_world.ecs_backend.addComponent(entity, SpriteAnimation{
